% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_sig.R
\name{add_sig}
\alias{add_sig}
\title{Add asterisks for presenting significance}
\usage{
add_sig(data)
}
\arguments{
\item{data}{a dataframe that has p.value}
}
\value{
data.frame object
}
\description{
Add asterisks for presenting significance
}
\keyword{add_sig}
