% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/basetable.R
\name{basetable}
\alias{basetable}
\title{Create a baseline characteristics table}
\usage{
basetable(
  data,
  varlist,
  grvar = NULL,
  len.max = 5,
  y.correct = T,
  trend = F,
  gr.name = NULL,
  method = 3,
  combine = F,
  print.result = T,
  p_norm = 0.05
)
}
\arguments{
\item{data}{Dataframe to be used}

\item{varlist}{Vector of variables to be displayed}

\item{grvar}{Grouping variable}

\item{len.max}{Maximum number of elements to convert to factor}

\item{y.correct}{a logical indicating whether to apply continuity correction}

\item{trend}{a logical whether to do trend test}

\item{gr.name}{Elements labels of grouping variables}

\item{method}{Select between parametric, nonparametric or combined testing}

\item{combine}{Display overall patient summary values as well}

\item{print.result}{Print varClassifier result}

\item{p_norm}{Cut-off p-value classifying normal and non-normal}
}
\value{
Dataframe of baseline characteristics table
}
\description{
This function create a baesline characteristics tables according to
the variables included in a specific dataframe.
}
\examples{
basetable(data=mtcars, varlist=c("cyl", "mpg", "wt", "hp"), grvar="am")

}
\keyword{basetable}
