% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ckd_epi.R
\name{ckd_epi}
\alias{ckd_epi}
\title{Calculate CKD-EPI eGFR}
\usage{
ckd_epi(Cr, sex, age, factor_male, factor_female, black = F, version = 2021)
}
\arguments{
\item{Cr}{serum creatine value}

\item{sex}{sex}

\item{age}{age}

\item{factor_male}{word refered to male}

\item{factor_female}{word refered to female}

\item{black}{black race or not}

\item{version}{CKD-EPI version (2009 or 2021)}
}
\value{
integer value
}
\description{
Calculate CKD-EPI eGFR
}
\keyword{ckd_epi}
