% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cow_plot.R
\name{cow_plot}
\alias{cow_plot}
\title{Creation of the cowplot object}
\usage{
cow_plot(
  n,
  ncol = n,
  font_size = 10,
  scale = 0.95,
  rel_widths = 1,
  uppercase = T,
  labels = NULL
)
}
\arguments{
\item{data}{a dataframe to be used}

\item{outcome}{a outcome variable}

\item{varlist}{predictors}

\item{round}{round value}

\item{print.result}{print the name of significant vairables}

\item{p.cut}{P-value cut-off}
}
\value{
cowplot obejct
}
\description{
Creation of the cowplot object
}
\examples{

}
\keyword{cow_plot}
