% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cox_iter.R
\name{cox_iter_cat}
\alias{cox_iter_cat}
\title{Making Cox PH models of multiple outcomes (categorical predictor variable)}
\usage{
cox_iter_cat(
  data,
  outcome_list,
  main_predictor,
  adjusted_var = NULL,
  round = 4
)
}
\arguments{
\item{data}{a dataframe to be used}

\item{outcome_list}{list of outcomes}

\item{main_predictor}{main predictor variable}

\item{adjusted_var}{variables list for adjustment}

\item{round}{round}
}
\value{
data.frame object
}
\description{
Making Cox PH models of multiple outcomes (categorical predictor variable)
}
\examples{

}
\keyword{cox_iter_cat}
