% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cox_uni.R
\name{cox_uni}
\alias{cox_uni}
\title{Automation of univariable Cox PH analysis}
\usage{
cox_uni(
  data,
  event,
  eventtime,
  vars,
  p.cut = 0.05,
  round = 3,
  digits = 1,
  nsmall = 3,
  print.result = F
)
}
\arguments{
\item{data}{a dataframe to be used}

\item{event}{a event variable}

\item{eventtime}{eventtime variable}

\item{vars}{predictors}

\item{p.cut}{P-value cut-off}

\item{round}{round value}

\item{digits}{digits}

\item{nsmall}{nsmall}

\item{print.result}{print the name of significant vairables}
}
\value{
data.frame
}
\description{
Automation of univariable Cox PH analysis
}
\keyword{cox_uni}
