% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_summary.R
\name{data_summary}
\alias{data_summary}
\title{Print multiple comparison results}
\usage{
data_summary(
  data,
  x,
  y,
  digits = 2,
  nsmall = 1,
  p.adj = "bonferroni",
  method = 2,
  print.option = c(1, 2, 3)
)
}
\arguments{
\item{data}{a dataframe to be used}

\item{x}{a variable for x-axis}

\item{y}{a variable for y-axis}

\item{digits}{digits}

\item{nsmall}{nsmall}

\item{p.adj}{P-value adjustment method for multiple comparison}

\item{method}{parametric (1) or non-parametric (2) test}

\item{print.option}{the options of print results}
}
\value{
print output
}
\description{
Print multiple comparison results
}
\examples{

}
\keyword{roc_comparison}
