% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggroc.R
\name{ggrocs}
\alias{ggrocs}
\title{Draw multiple ROC plots}
\usage{
ggrocs(
  roclist,
  showAUC = TRUE,
  interval = 0.2,
  breaks = seq(0, 1, interval),
  labels = NULL
)
}
\arguments{
\item{roclist}{a list of roc objects}

\item{showAUC}{show AUC value}

\item{interval}{interval}

\item{breaks}{breaks}

\item{labels}{label of ROC lines}
}
\value{
ggplot object
}
\description{
Draw multiple ROC plots
}
\keyword{ggrocs}
