% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hist_group.R
\name{hist_group}
\alias{hist_group}
\title{Make a histogram categorized by grouping variable}
\usage{
hist_group(
  data,
  var,
  group,
  show.hist = T,
  show.density = T,
  show.mean = T,
  bins = 50,
  alpha = 0.5
)
}
\arguments{
\item{data}{a dataframe to be used}

\item{var}{a numeric variable}

\item{group}{a grouping variable (factor)}

\item{show.hist}{a logical indicating whether to show histogram}

\item{show.density}{a logical indicating whether to show density plot}

\item{bins}{number of bins}

\item{alpha}{transparency of density plot}
}
\value{
ggplot2 histogram
}
\description{
This function make a histogram categorized by grouping variable
}
\examples{
hist_group(mtcars, "mpg", "am")

}
\keyword{hist_group}
