% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iter_cor.R
\name{iter_cor}
\alias{iter_cor}
\title{Iteration of correlation analyses with a specific variable}
\usage{
iter_cor(data, varlist, var, method = "pearson", print.result = F, p.round = 3)
}
\arguments{
\item{data}{a dataframe to be used}

\item{varlist}{a vector of continuous variables}

\item{var}{a variable of interest}

\item{method}{a character string indicating which correlation coefficient is to be used for the test. One of "pearson", "kendall", or "spearman", can be abbreviated.}

\item{print.result}{a logical indicating whether to show significant variables}

\item{p.round}{a specified number of decimal places to which rounds the values}
}
\value{
Dataframe containing the result of correlation analyes
}
\description{
This function perform multiple correlation analyses between a specific variable
and various continuous variables
}
\examples{
iter_cor(data=mtcars, varlist=c("cyl", "wt", "hp"), var="mpg", print.result=T, p.round=4)

}
\keyword{iter_cor}
