% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iter_t.R
\name{iter_t}
\alias{iter_t}
\title{Iteration of Student's t-tests with a specific variable}
\usage{
iter_t(data, varlist, var, method = 1, maxgroup = 5, p.round = 3)
}
\arguments{
\item{data}{a dataframe to be used}

\item{varlist}{a vector of categorical variables}

\item{var}{a variable of interest}

\item{method}{parametric (1) or non-parametric test (2)}

\item{maxgroup}{a maximum number of groups}

\item{p.round}{a specified number of decimal places to which rounds the values}
}
\value{
Dataframe containing the result of t-tests
}
\description{
This function perform Student's t-tests between a specific variable
and various categorical variables
}
\keyword{iter_t}
