% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loadpacks.R
\name{loadpacks}
\alias{loadpacks}
\title{Load R packages}
\usage{
loadpacks(pkg, ...)
}
\arguments{
\item{pkg}{R packages to be loaded}

\item{...}{...}
}
\description{
This function loads R packages easily.
}
\keyword{loadpacks}
