% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logistic_models_val.R
\name{logistic_models_val}
\alias{logistic_models_val}
\title{Validation of logistic regression results of multiple models}
\usage{
logistic_models_val(
  data,
  outcome,
  models,
  n.aic = 10,
  n.bic = 10,
  fold = 10,
  repeats = 100,
  boot.repeat = 1000
)
}
\arguments{
\item{data}{a dataframe to be used}

\item{outcome}{binary outcome variable}

\item{models}{logistic regression models}

\item{n.aic}{number of the combination to be showed (arranged by AIC values)}

\item{n.bic}{number of the combination to be showed (arranged by BIC values)}

\item{fold}{cross-validation fold number}

\item{repeats}{cross-validation repeat number}

\item{boot.repeat}{boot-strap repeat number}
}
\value{
data.frame object
}
\description{
Validation of logistic regression results of multiple models
}
\keyword{logistic_models_val}
