% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logistic_uni.R
\name{logistic_uni}
\alias{logistic_uni}
\title{Automation of univariable logistic analysis}
\usage{
logistic_uni(data, outcome, varlist, round = 3, print.result = T, p.cut = 0.05)
}
\arguments{
\item{data}{a dataframe to be used}

\item{outcome}{a outcome variable}

\item{varlist}{predictors}

\item{round}{round value}

\item{print.result}{print the name of significant vairables}

\item{p.cut}{P-value cut-off}
}
\value{
data.frame
}
\description{
Automation of univariable logistic analysis
}
\keyword{logistic_uni}
