% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logistic_val.R
\name{logistic_val}
\alias{logistic_val}
\title{Validation of logistic regression results with the combiation of variables}
\usage{
logistic_val(
  data,
  outcome,
  vars,
  auc.cutoff = 0.65,
  n.aic = 10,
  n.bic = 10,
  fold = 10,
  repeats = 100,
  boot.repeat = 1000,
  maxvar = ceiling(min(table(data[[outcome]])) * 0.1) + 2
)
}
\arguments{
\item{data}{a dataframe to be used}

\item{outcome}{binary outcome variable}

\item{vars}{predictor variables}

\item{auc.cutoff}{cut-off value of AUC}

\item{n.aic}{number of the combination to be showed (arranged by AIC values)}

\item{n.bic}{number of the combination to be showed (arranged by BIC values)}

\item{fold}{cross-validation fold number}

\item{repeats}{cross-validation repeat number}

\item{boot.repeat}{boot-strap repeat number}

\item{maxvar}{maximum number of variables}
}
\value{
data.frame object
}
\description{
Validation of logistic regression results with the combiation of variables
}
\examples{

}
\keyword{roc_comparison}
