% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/newid_exchanger.R
\name{newid_exchanger}
\alias{newid_exchanger}
\title{Exchange old patient ID to new study ID}
\usage{
newid_exchanger(data, newid_data, old_id_col, new_id_col = "studyid")
}
\arguments{
\item{data}{original data}

\item{newid_data}{dataframe with old ID and new ID}

\item{old_id_col}{column pointing to old ID}

\item{new_id_col}{study ID column}
}
\value{
data.frame object
}
\description{
Exchange old patient ID to new study ID
}
\keyword{newid_exchanger}
