% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/newid_gen.R
\name{newid_gen}
\alias{newid_gen}
\title{New study ID generator}
\usage{
newid_gen(seed, prefix = "", max_digit = 7)
}
\arguments{
\item{seed}{seed number}

\item{prefix}{prefix for study ID}

\item{max_digit}{maximum digits}
}
\value{
data.frame
}
\description{
New study ID generator
}
\keyword{newid_gen}
