% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_bee.R
\name{plot_bee}
\alias{plot_bee}
\title{Beeswarm plot with confidence intervals}
\usage{
plot_bee(
  data,
  x,
  y,
  bar.width = 0.5,
  err.width = 0.15,
  xlab = x,
  ylab = y,
  xlabels = levels(data[[x]]),
  se = F,
  cex = 2.5,
  dot = T,
  shape = rep(16, 10),
  bar_graph = F
)
}
\arguments{
\item{data}{a dataframe to be used}

\item{x}{a variable for x-axis}

\item{y}{a variable for y-axis}

\item{bar.width}{mean bar width}

\item{err.width}{error bar width}

\item{xlab}{x-axis label}

\item{ylab}{y-axis label}

\item{xlabels}{labels of x-axis groups}

\item{se}{SE (T) or 95\% CI (F)}

\item{cex}{size of the dots}

\item{dot}{revealing dots or not}

\item{shape}{shape of the dots}

\item{bar_graph}{bar_graph (T) or beeswarm graph (F)}
}
\value{
ggplot object
}
\description{
Draw a beeswarm plot
}
\keyword{plot_bee}
