% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_cor.R
\name{plot_cor}
\alias{plot_cor}
\title{Correlation plot}
\usage{
plot_cor(
  data,
  x,
  y,
  xlab = x,
  ylab = y,
  r = T,
  p = T,
  xpos = 0.7,
  ypos = 0.7,
  x2pos = xpos,
  y2pos = ypos - 0.1,
  method = "pearson",
  round = 3,
  rsize = 4,
  low.limit = F,
  ci = T
)
}
\arguments{
\item{data}{a dataframe to be used}

\item{x}{a variable for x-axis}

\item{y}{a variable for y-axis}

\item{r}{correlation coefficient}

\item{p}{P-value}

\item{xpos}{relative location of correlation coefficient}

\item{ypos}{relative location of correlation coefficient}

\item{x2pos}{relative x location of P-value}

\item{y2pos}{relative y location of P-value}

\item{method}{Pearson or Spearman test}

\item{round}{round}

\item{rsize}{rsize}

\item{low.limit}{low.limit}

\item{ci}{confidence interval}
}
\value{
ggplot object
}
\description{
Draw a correlation plot
}
\keyword{plot_cor}
