% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_csvk.R
\name{read_csvk}
\alias{read_csvk}
\title{Read CSV file with EUC-KR encoding}
\usage{
read_csvk(file, ...)
}
\arguments{
\item{file}{CSV file}

\item{...}{parameters of read_csv}
}
\value{
data.frame object
}
\description{
Read CSV file with EUC-KR encoding
}
\keyword{read_csvk}
