% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/roc_comparison.R
\name{roc_comparison}
\alias{roc_comparison}
\title{Comparison of two ROC objects}
\usage{
roc_comparison(roc1, roc2, venkatraman = F)
}
\arguments{
\item{roc1}{roc object}

\item{roc2}{roc object}

\item{venkatraman}{add the result of venkatraman method}
}
\value{
data.frame
}
\description{
Comparison of two ROC objects
}
\keyword{roc_comparison}
