% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/basetable.R
\name{varClassifier}
\alias{varClassifier}
\title{Classify variables to categorical or numerical variables}
\usage{
varClassifier(dt, myVars, len.max = 5, print.result = T, p_norm = 0.05)
}
\arguments{
\item{dt}{Dataframe to be used}

\item{myVars}{Variables to be categorized}

\item{len.max}{Maximum number of elements to convert to factor}

\item{print.result}{Print the analyzed result}

\item{p_norm}{Cut-off p-value classifying normal and non-normal}
}
\value{
lists of three different types
}
\description{
This function classifies variables to three differente types including
categorical, non-normal numerical or normal numerical variables.
}
\examples{
varClassifier(dt=mtcars, myVars=c("am", "cyl", "mpg"))

}
\keyword{varClassifier}
