% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vec.R
\name{vec}
\alias{vec}
\title{Make words to vector}
\usage{
vec(vars, ...)
}
\arguments{
\item{vars}{words to be a vector}
}
\value{
vector
}
\description{
This function makes a vector from words list
}
\examples{
vec(a, b, c)

}
\keyword{vec}
