% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/x_table.R
\name{x_table}
\alias{x_table}
\title{x by x table (n & percent)}
\usage{
x_table(data, outcome, predictor)
}
\arguments{
\item{data}{a dataframe to be used}

\item{outcome}{outcome variable}
}
\value{
data.frame object
}
\description{
x by x table (n & percent)
}
\examples{

}
\keyword{x_table}
