% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cow_plot.R
\name{cow_plot}
\alias{cow_plot}
\title{Creation of the cowplot object}
\usage{
cow_plot(
  n,
  ncol = n,
  font_size = 10,
  scale = 0.95,
  rel_widths = 1,
  uppercase = T,
  labels = NULL
)
}
\arguments{
\item{n}{a number of plots}

\item{ncol}{a number of columns}

\item{font_size}{font size}

\item{scale}{scale}

\item{rel_widths}{relative width of columns}

\item{uppercase}{uppercase}

\item{labels}{labels}
}
\value{
cowplot obejct
}
\description{
Creation of the cowplot object
}
\keyword{cow_plot}
